stored credentials
Username and password are now saved under single string name. as in "username:password"  the string name is "credentials"
