+++ blackJack/src/BlackjackGUI.java
@@ -1,88 +1,161 @@
 
 import javax.swing.*;
 import java.awt.*;
-import java.io.*;
-import java.net.Socket;
 
 public class BlackjackGUI {
-    private JFrame frame; // Main window frame for the GUI
+    private JFrame frame;
     private CardLayout cardLayout;
     private JPanel cardPanel;
-    private JTextField usernameField;
-    private JPasswordField passwordField;
-    private Socket socket;
-    private ObjectOutputStream outputStream;
-    private ObjectInputStream inputStream;
+    private String iconPath = "C:/Users/Zayyyy/Documents/CS401_BlackJack1/CS401-Blackjack/project pictures/icon.png"; // Path to the icon image
 
-    // Constructor
     public BlackjackGUI() {
-        initializeGUI(); // Initialize GUI components when the object is created
+        initializeGUI();
     }
 
-	// Method to initialize the whole GUI
     private void initializeGUI() {
-        frame = new JFrame("Blackjack Casino Game"); // Create the frame with a title
-        frame.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE); // Set default close operation
-        frame.setSize(800, 600); // Set size of the frame
-        
-        // CardLayout allows switching between different panels (like login and game panels)
-        CardLayout cardLayout = new CardLayout();
-        JPanel cardPanel = new JPanel(cardLayout); // Panel that holds the different screens
+        frame = new JFrame("BLACKJACK");
+        frame.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
+        frame.setSize(800, 600);
 
-        // Login Panel Setup
-        JPanel loginPanel = new JPanel(new GridLayout(3, 2)); // Grid layout for orderly inputs
-        JLabel usernameLabel = new JLabel("Username:"); // Label for username
-        JTextField usernameField = new JTextField(); // Text field for username input
-        JLabel passwordLabel = new JLabel("Password:"); // Label for password
-        JPasswordField passwordField = new JPasswordField(); // Password field for secure password input
-        JButton loginButton = new JButton("Login"); // Login button
-        JButton registerButton = new JButton("Register"); // Register button
+        cardLayout = new CardLayout();
+        cardPanel = new JPanel(cardLayout);
+        cardPanel.setBackground(new Color(32, 32, 32)); // A dark background for the card panel
 
-        // Add components to the login panel
-        loginPanel.add(usernameLabel);
-        loginPanel.add(usernameField);
-        loginPanel.add(passwordLabel);
-        loginPanel.add(passwordField);
-        loginPanel.add(loginButton);
-        loginPanel.add(registerButton);
-
-        // Main Game Panel
-        JPanel gamePanel = new JPanel(); // Panel for the game interface
-        gamePanel.add(new JLabel("Welcome to the Blackjack Table!")); // Welcome label
-
-        // Add both panels to the card panel with identifiers
-        cardPanel.add(loginPanel, "Login");
-        cardPanel.add(gamePanel, "Game");
-
-        // Action Listeners for buttons
-        loginButton.addActionListener(e -> sendLogin(usernameField.getText(), new String(passwordField.getPassword())));
-        registerButton.addActionListener(e -> JOptionPane.showMessageDialog(frame, "Registration Module not implemented."));
+        initializeLoginPanel();
+        initializeGamePanel();
+        initializeBlackjackTablePanel();
 
         frame.add(cardPanel);
         frame.setVisible(true);
     }
 
-    // Handle login by sending a message to the server and processing the response.
-    private void sendLogin(String username, String password) {
-        try {
-            Message loginMessage = new Message(Type.Login, Status.New, username + ":" + password);
-            outputStream.writeObject(loginMessage);
-            //outputStream.flush();
-            Message response = (Message) inputStream.readObject();
-            if (response.getStatus() == Status.Success) {
-                JOptionPane.showMessageDialog(frame, "Login Successful");
-                cardLayout.show(cardPanel, "Game");
-            } else {
-                JOptionPane.showMessageDialog(frame, "Login Failed");
-            }
-        } catch (IOException | ClassNotFoundException e) {
-            JOptionPane.showMessageDialog(frame, "Error communicating with server");
-            e.printStackTrace();
+    private void initializeLoginPanel() {
+        JPanel loginPanel = new JPanel(new GridBagLayout());
+        loginPanel.setBackground(new Color(32, 32, 32)); // Dark background
+        GridBagConstraints gbc = new GridBagConstraints();
+
+        // Adding the icon to the side of the login panel
+        ImageIcon icon = new ImageIcon(iconPath);
+        JLabel iconLabel = new JLabel(icon);
+        gbc.gridx = 0;
+        gbc.gridy = 0;
+        gbc.gridheight = GridBagConstraints.REMAINDER;
+        gbc.anchor = GridBagConstraints.NORTH;
+        gbc.insets = new Insets(10, 10, 10, 10);
+        loginPanel.add(iconLabel, gbc);
+
+        // Resetting constraints for other components
+        gbc.gridheight = 1;
+        gbc.anchor = GridBagConstraints.CENTER;
+
+        JLabel welcomeLabel = new JLabel("WELCOME");
+        welcomeLabel.setForeground(Color.WHITE);
+        welcomeLabel.setFont(new Font("Arial", Font.BOLD, 24));
+        gbc.gridx = 1;
+        gbc.insets = new Insets(10, 0, 20, 0);
+        loginPanel.add(welcomeLabel, gbc);
+
+        JLabel signInLabel = new JLabel("Sign In");
+        signInLabel.setForeground(Color.WHITE);
+        signInLabel.setFont(new Font("Arial", Font.BOLD, 18));
+        gbc.gridy++;
+        loginPanel.add(signInLabel, gbc);
+
+        JTextField usernameField = new JTextField(20);
+        customizeComponent(usernameField);
+        gbc.gridy++;
+        loginPanel.add(usernameField, gbc);
+
+        JPasswordField passwordField = new JPasswordField(20);
+        customizeComponent(passwordField);
+        gbc.gridy++;
+        loginPanel.add(passwordField, gbc);
+
+        JButton loginButton = new JButton("LOGIN");
+        customizeButton(loginButton);
+        gbc.gridy++;
+        loginPanel.add(loginButton, gbc);
+        
+        loginButton.addActionListener(e -> {
+            String username = usernameField.getText();  // Fetch the username entered
+            String password = new String(passwordField.getPassword());  // Fetch the password entered
+            String credentials = username + ":" + password;  // Combine username and password with a colon separator
+            System.out.println("Combined credentials: " + credentials);  // Optionally, print the combined credentials
+            cardLayout.show(cardPanel, "Game");  // Move to the game panel
+        });
+
+        JButton registerButton = new JButton("SIGN UP NOW");
+        customizeButton(registerButton);
+        gbc.gridy++;
+        loginPanel.add(registerButton, gbc);
+
+        cardPanel.add(loginPanel, "Login");
+        loginButton.addActionListener(e -> cardLayout.show(cardPanel, "Game"));
+        registerButton.addActionListener(e -> JOptionPane.showMessageDialog(frame, "Registration not implemented"));
+    }
+
+    private void initializeGamePanel() {
+        // Custom game panel - This will be your actual game interface
+        JPanel gamePanel = new JPanel(new GridBagLayout());
+        gamePanel.setBackground(new Color(0, 102, 0)); // Set to a green resembling a blackjack table
+
+        GridBagConstraints gbcGame = new GridBagConstraints();
+        gbcGame.gridwidth = GridBagConstraints.REMAINDER;
+        gbcGame.anchor = GridBagConstraints.CENTER;
+        gbcGame.fill = GridBagConstraints.HORIZONTAL;
+
+        JLabel welcomeToBlackjack = new JLabel("Welcome to BLACKJACK!", SwingConstants.CENTER);
+        welcomeToBlackjack.setForeground(Color.WHITE); // White text
+        welcomeToBlackjack.setFont(new Font("Arial", Font.BOLD, 40)); // Larger font size for the title
+        gbcGame.insets = new Insets(20, 0, 20, 0);
+        gamePanel.add(welcomeToBlackjack, gbcGame);
+
+        JButton playButton = new JButton("PLAY");
+        customizeButton(playButton);
+        playButton.addActionListener(e -> cardLayout.show(cardPanel, "BlackjackTable"));
+        gbcGame.insets = new Insets(10, 0, 10, 0);
+        gamePanel.add(playButton, gbcGame);
+
+        JButton exitButton = new JButton("EXIT");
+        customizeButton(exitButton);
+        exitButton.addActionListener(e -> frame.dispose());
+        gamePanel.add(exitButton, gbcGame);
+
+        JLabel footerLabel = new JLabel("This game is brought to you by Group 5", SwingConstants.CENTER);
+        footerLabel.setForeground(Color.WHITE);
+        footerLabel.setFont(new Font("Arial", Font.PLAIN, 16));
+        gbcGame.insets = new Insets(40, 0, 0, 0);
+        gamePanel.add(footerLabel, gbcGame);
+
+        cardPanel.add(gamePanel, "Game");
+    }
+
+    private void initializeBlackjackTablePanel() {
+        // Placeholder for the actual blackjack table panel
+        JPanel blackjackTablePanel = new JPanel();
+        blackjackTablePanel.setBackground(new Color(0, 102, 0)); // Green table background
+        // TODO: Add actual game components here
+
+        cardPanel.add(blackjackTablePanel, "BlackjackTable");
+    }
+
+    private void customizeComponent(JComponent component) {
+        component.setPreferredSize(new Dimension(150, 30));
+        component.setFont(new Font("Arial", Font.PLAIN, 14));
+        component.setForeground(Color.WHITE);
+        component.setBackground(new Color(64, 64, 64));
+        if (component instanceof JTextField) {
+            ((JTextField) component).setCaretColor(Color.WHITE);
         }
     }
 
-    // Main method to run the application.
+    private void customizeButton(JButton button) {
+        button.setFont(new Font("Arial", Font.BOLD, 14));
+        button.setBackground(new Color(0, 123, 255));
+        button.setForeground(Color.WHITE);
+    }
+
     public static void main(String[] args) {
-        SwingUtilities.invokeLater(BlackjackGUI::new);
+        SwingUtilities.invokeLater(GUITest::new);
     }
 }
